Rename netflix_light to netflix_viewing_history
