+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@corsali/userdata-extractor",
-  "version": "0.3.0-beta6",
+  "version": "0.3.0",
   "author": "Vana",
   "type": "module",
   "scripts": {
