--- src/extractor/netflixLight/models/netflixLightBaseModel.ts
@@ -1,3 +1,0 @@
-import { TableRow } from "../../../models/table/tableRow.js";
-
-export class NetflixLightBaseModel extends TableRow {}
