+++ src/extractor/netflixViewingHistory/models/netflixViewingHistoryBaseModel.ts
@@ -1,0 +1,3 @@
+import { TableRow } from "../../../models/table/tableRow.js";
+
+export class NetflixViewingHistoryBaseModel extends TableRow {}
