+++ test/utils/fileToZip.test.ts
@@ -19,22 +19,25 @@ describe("utils/fileToZip", () => {
 
   it("fileToZip() returns a ServiceFile", async () => {
     deleteFile(
-      `csv/${config.SERVICE_NETFLIX_LIGHT}/${config.SERVICE_NETFLIX_LIGHT}.sqlite`
+      `csv/${config.SERVICE_NETFLIX_VIEWING_HISTORY}/${config.SERVICE_NETFLIX_VIEWING_HISTORY}.sqlite`
     );
     const file = await loadTestFile(
-      `csv/${config.SERVICE_NETFLIX_LIGHT}/NetflixViewingHistory.csv`
+      `csv/${config.SERVICE_NETFLIX_VIEWING_HISTORY}/NetflixViewingHistory.csv`
     );
-    const serviceFile = await filesToZip(config.SERVICE_NETFLIX_LIGHT, [file]);
+    const serviceFile = await filesToZip(
+      config.SERVICE_NETFLIX_VIEWING_HISTORY,
+      [file]
+    );
     const database = await zipToSQLiteInstance([serviceFile], false);
 
     saveSqliteDump(
-      database.exportDatabase(config.SERVICE_NETFLIX_LIGHT),
-      `csv/${config.SERVICE_NETFLIX_LIGHT}/${config.SERVICE_NETFLIX_LIGHT}.sqlite`
+      database.exportDatabase(config.SERVICE_NETFLIX_VIEWING_HISTORY),
+      `csv/${config.SERVICE_NETFLIX_VIEWING_HISTORY}/${config.SERVICE_NETFLIX_VIEWING_HISTORY}.sqlite`
     );
     database.close();
     expect(
       fileExists(
-        `csv/${config.SERVICE_NETFLIX_LIGHT}/${config.SERVICE_NETFLIX_LIGHT}.sqlite`
+        `csv/${config.SERVICE_NETFLIX_VIEWING_HISTORY}/${config.SERVICE_NETFLIX_VIEWING_HISTORY}.sqlite`
       )
     ).toBeTruthy();
   });
